requirements.txt
feedparser==6.0.11
python-dotenv==1.0.0
APScheduler==3.10.4
PyYAML==6.0.1
